Bug-fix: resolve outdated entities (#2)
* Refactor entity classes for improved readability and consistency

* Add AdminRepository and update ApplicationInitConfig for admin user creation

* Resolve User-related classes according to latest changes in database, update user identification to use email

* Update authentication to use email instead of username for user identification

* Change ID types from Long to Integer in entity classes and update UserController methods accordingly

+++ src/main/java/com/visita/entities/BookingEntity.java
@@ -1,11 +1,26 @@
 package com.visita.entities;
 
-import jakarta.persistence.*;
-import lombok.*;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "bookings")
@@ -15,45 +30,45 @@ import java.util.List;
 @Builder
 public class BookingEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "booking_id")
-    private Long bookingId;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "booking_id")
+	private Integer bookingId;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private UserEntity user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private UserEntity user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "tour_id")
-    private TourEntity tour;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "tour_id")
+	private TourEntity tour;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "promotion_id")
-    private PromotionEntity promotion;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "promotion_id")
+	private PromotionEntity promotion;
 
-    @Column(name = "booking_date")
-    private LocalDateTime bookingDate;
+	@Column(name = "booking_date")
+	private LocalDateTime bookingDate;
 
-    @Column(name = "num_adults", nullable = false)
-    private Integer numAdults;
+	@Column(name = "num_adults", nullable = false)
+	private Integer numAdults;
 
-    @Column(name = "num_children")
-    private Integer numChildren;
+	@Column(name = "num_children")
+	private Integer numChildren;
 
-    @Column(name = "total_price", nullable = false, precision = 15, scale = 2)
-    private BigDecimal totalPrice;
+	@Column(name = "total_price", nullable = false, precision = 15, scale = 2)
+	private BigDecimal totalPrice;
 
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "ENUM('Pending','Confirmed','Cancelled','Completed')")
-    private BookingStatus status;
+	@Enumerated(EnumType.STRING)
+	@Column(columnDefinition = "ENUM('Pending','Confirmed','Cancelled','Completed')")
+	private BookingStatus status;
 
-    @Column(name = "special_request", length = 500)
-    private String specialRequest;
+	@Column(name = "special_request", length = 500)
+	private String specialRequest;
 
-    @OneToMany(mappedBy = "booking", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private List<InvoiceEntity> invoices;
+	@OneToMany(mappedBy = "booking", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	private List<InvoiceEntity> invoices;
 
-    @OneToMany(mappedBy = "booking", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private List<PaymentEntity> payments;
+	@OneToMany(mappedBy = "booking", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	private List<PaymentEntity> payments;
 }
